Range analysis

+++ ext/mockturtle/include/mockturtle/algorithms/dsd_decomposition.hpp
@@ -68,7 +68,7 @@ public:
         _on_prime( on_prime ),
         _ps( ps )
   {
-    for ( auto i = 0; i < func.num_vars(); ++i )
+    for ( auto i = 0u; i < func.num_vars(); ++i )
     {
       if ( kitty::has_var( func, i ) )
       {
@@ -184,7 +184,7 @@ public:
     }
     auto prime_large = remainder;
     kitty::min_base_inplace( prime_large );
-    auto prime = kitty::shrink_to( prime_large, support.size() );
+    auto prime = kitty::shrink_to( prime_large, static_cast<unsigned int>( support.size() ) );
     return _on_prime( prime, new_pis );
   }
